Un-stub shiny rate editor
Precomputed a list of instructions (format u16 rerollCount, u16 half of
instruction); upper16 bits are constant (A0 E3).

If a reroll count that is not in the precomputed list, pk3DS will bump
the rate up to the nearest precomputed value.

Increased the default value to 125 to make it obvious that 3% was not
the existing rate, and expanded the label usage. With so much text on
the form the actual editing area is now indicated by the red text.